Simplify IDisposable object returned from Subscribe.
