+++ src/Sarif.Viewer.VisualStudio/Sarif.Viewer.VisualStudio.csproj
@@ -183,7 +183,7 @@
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Models\SarifErrorListItem.cs" />
     <Compile Include="SdkUIUtilities.cs" />
-    <Compile Include="ErrorList\SarifTableDataSink.cs" />
+    <Compile Include="ErrorList\SinkWrapper.cs" />
     <Compile Include="ErrorList\SarifTableDataSource.cs" />
     <Compile Include="ViewModels\ViewModelLocator.cs" />
     <Compile Include="WeakEventHandlerManager.cs" />
